pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.org"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.3"
sqlmodel = "^0.0.22"
fastapi = {extras = ["standard"], version = "^0.115.3"}
alembic = "^1.13.3"
cryptography = "^43.0.3"
pydantic-settings = "^2.6.0"
psycopg = "^3.2.3"
psycopg-binary = "^3.2.3"
passlib = "^1.7.4"
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
